net: split out functions related to registering inflight socket files
commit f4e65870e5cede5ca1ec0006b6c9803994e5f7b8 upstream.

We need this functionality for the io_uring file registration, but
we cannot rely on it since CONFIG_UNIX can be modular. Move the helpers
to a separate file, that's always builtin to the kernel if CONFIG_UNIX is
m/y.

No functional changes in this patch, just moving code around.

Reviewed-by: Hannes Reinecke <hare@suse.com>
Acked-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
[ backported to older kernels to get access to unix_gc_lock - gregkh ]
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Change-Id: I0ea543fef5b17e97e12182ab5c62bd6622a6c85b
Git-commit: eee65a1282369eedfcbb664d0c865a0ef3eb7017
Git-repo: https://android.googlesource.com/kernel/msm
Signed-off-by: PavanKumar S.R <quic_pavasr@quicinc.com>